Move CommandManager
