+++ src/nl/hannahsten/texifyidea/lang/alias/AliasManager.kt
@@ -1,0 +1,7 @@
+package nl.hannahsten.texifyidea.lang.alias
+
+/**
+ * Manage aliases.
+ */
+class AliasManager {
+}
