+++ src/nl/hannahsten/texifyidea/lang/alias/EnvironmentManager.kt
@@ -1,0 +1,7 @@
+package nl.hannahsten.texifyidea.lang.alias
+
+/**
+ * Similar to the [CommandManager], this manages aliases of environments.
+ */
+class EnvironmentManager {
+}
